Add option for tracing AIDL calls.
This allows generating AIDL code that traces to the newly
introduced ATRACE_TAG_AIDL, for both c++ and Java targets.

The "::client" traces wrap the complete call, from entering
the proxy right until returning the result. As such, this
tag can be used to measure the complete duration of the IPC
call.

The "::server" traces wrap the actual execution of the
method on the server side, so it can be used to determine
the actual time the server took to complete the work.

Subtracting the "::server" slice from the corresponding
"::client" slice allows determining the overhead binder
itself introduced.

This is similar to how HIDL calls are traced.

Bug: 74416314
Test: aidl_unittest passed (incl new tests),
      verified atrace outpout
Change-Id: Id9a0207d53ebbdb6a21d8d96f56b8eee116bd477
Merged-In: Id9a0207d53ebbdb6a21d8d96f56b8eee116bd477