Enable glr parser and set rr 0.
As a side-effect, annotations are disabled for
structured parcelables (which aren't used anywhere).

They will be re-enabled as part of b/110967839, b/70046217
when I restructure all declarations to be in the same scope.

The parser needs to be set to glr in order to restrict
the number of reduce/reduce conflicts to 0. We are not
actually parsing ambiguous grammars here currently.

Bug: 110758329
Test: runtests.sh
Change-Id: I2fd4571f6e2440efe068746746d1337d08c5dc9b