Type parameters are parsed by yacc.
This is the start of the attempt to re-write or clean type_namespace.h.
An AIDL type with generic type parameters, i.e., List<String>, is now
parsed by bison and the type parameters themselves are modeled as
AIDL types.

Note that nesting of type parameters is still not supported though. The
syntax allows it but type_namespace.h is preventing it, which will be
fixed in following CLs.

Bug: 110967839
Test: m -j
Test: system/tools/aidl/runtests.sh

Change-Id: Idc55ac4dde22ecc7377211a43a5a3759c0f25923