add --structured flag
Before, using an aidl_interface with unstructured parcelables
would cause Java/C++ specific errors.

Now, you get a nice error message:
android.aidl.tests.SimpleParcelable is not structured, but this is a structured interface

Bug: 111117220
Test: runtests.sh
Test: manually add --structured when compiling interface, and see errors for every violation:
aidl E 07-30 12:04:04 146895 146895 aidl.cpp:676] android.content.ComponentName is not structured, but this is a structured interface.
aidl E 07-30 12:04:04 146895 146895 aidl.cpp:676] android.content.Intent is not structured, but this is a structured interface.
aidl E 07-30 12:04:04 146895 146895 aidl.cpp:676] android.os.ParcelFileDescriptor is not structured, but this is a structured interface

Change-Id: I38c7567d731f89823b4afd07f29a07d285755424