add --version option to specify version to AIDL interface
When compiling an AIDL file, one can specify the version of the
interface in the AIDL file using --version option. When the option is
specified, the code is generated with methods/fields for querying the
interface version. Specifically,

the version of the interface that the caller is built against is
retrieved by reading a constant variable IFoo::VERSION (in C++) and
IFoo.VERSION (in Java), while the interface that the callee (the binder
service) is currently implementing is retrieved by calling
getInterfaceVersion().

For Java, when the generated classes (IFoo, IFoo.Stubs, etc.) are shared
in between the client and the server -- i.e. the classes are in the
bootclasspath --, the server is required to implement
getInterfaceVersion() as follows:

class MyFoo extends IFoo.Stubs {
    @Override
    public final int getInterfaceVersion() { return IFoo.VERSION; }

    ...
}

Since IFoo.VERSION is a 'static final int', the value is embedded into
overriding method and is not updated even when IFoo class is updated to
a new version.

For C++, this isn't required because the generated classes (BnInterface,
etc.) can't be shared across the client and server due to the lack of
ABI in C++.

Bug: 110760419
Test: m -j
Test: system/tools/aidl/runtest.sh
Change-Id: I65b538a3834316c3b694d3250999d2ffd38eb4c9