getInterfaceVersion() return -1 when remote is not implementing it
getInterfaceVersion() in the default impl class is returning -1 to
indicate that the remote side is not imeplementing the method, which
can't be happening unless the remote side was implemented without
using the automatically generated code. Since not implementing the meta
transaction is an error, returning -1 to indicate that. (0 means
'unspecified' which is for interfaces without version specified.

Bug: 110760419
Test: m -j
Test: runtests.sh

Change-Id: If0bcfdd295f3e85a18430111821c8c7d9a7939c7