--header_out directory is automatically created if not exist
Just like the --out option, --header_out DIR creates directory DIR when
it does not exist.

Also this CL ensures that directories are always ends with the path
separator inside the AIDL compiler. This is not only for consistency
but also to make it easy to determine whether a path (which hasn't yet
been created) is a dir or a file.

Bug: 110967839
Test: runtests.sh
Test: aidl --lang=cpp -o ~/temp/out -h ~/temp/header
~/temp/foo/bar/ITest.aidl
(both ~/temp/out and ~/temp/header are non-existing)
is successful

Change-Id: I48b4c7fe1334651a21c0ff1fa6ab6451a6128b1b