Don't resolve symbols for --checkapi
Type specifiers in an AIDL API file (generated by `aidl --dumpapi`) are
all fully qualified. We don't need to run the resolution then.

This is also required because types that are not part of the API (e.g.,
a type defined in other AIDL interface) are not imported. Nothing is
imported in an API file because we are NOT interested in the API
specification of the imported type; we are only interested in the APIs
of the types in 'this' interface. Due to the lack of importing, the
resolution fails for the imported types. So skipping them only for
--checkapi.

Bug: 110758635
Test: aidl --checkapi build/api/test-piece-2/current.aidl build/api/test-piece-2/current.aidl && echo success
shows 'success'. The API file has reference to 'some_package.IFoo' which
isn't part of the API.

Change-Id: Ie4bdf0435a953a7a6e84d88d5d53e842cb9363df