api check is done whenever an aidl_interface is build
API check for AIDL interface is done whenever the module is built. This
ensures that the API check is performed even when the module is built
via m <module_name> or mma under a directory.

This also changes the names of the phony targets for api checking and
updated in order to be more readable. (e.g., checkaidlapi ->
check-aidl-api)

Bug: 113133819
Test: m -j test-piece-1-cpp fails when api/test-piece-1/curret.aidl is
changed

Change-Id: Ia85d0410dd2846e08d434558e5742c8bb23b3f1c