Allow Stable AIDL to have manually assigned IDs
It is possible to have a stable AIDL interface without fixing the IDs
because they are assigned IDs by texture order and the order is checked
when duing the api check.

However, there are some AIDL interfaces that already have manually
assigned IDs. In order to ease the migration of the interfaces to the
stable AIDL, it is now allowed for a stable AIDL to have manually
assigned IDs. In this case, the IDs are emitted to the API dump.

Bug: 110758635
Test: runtests.sh
Change-Id: I9186a4daae726b9dd80308844690af7538e8d1ba