Unstructured parcelables: behavior for many files.
Before this CL, if you had an interface and a parcelable in the same
file, you would hit a check.

Now, based on -b (which is set in the Android build system due to its
dependency requirements) unstructured parcelables are either allowed
or disallowed when compiling multiple files. Meaning, in Android
platform builds, the correct error is shown, and people using other
build systems which depend on the default behavior can compile AIDL
files with parcelables. As to avoid restricting AIDL in the future,
this will likely be cleaned up in b/113546226. We have no current
usecase for allowing unstructured parcelables in files containing
multiple definitions, but for now, it as at least consistent.

Bug: 112017361
Test: runtests.sh (tests added for this)
Change-Id: I533a4995d2c4ea69bf7f331f357f8667694fb8e4