Add versions property to aidl_interface
"versions" property is the list of version names of an AIDL under the
directory specified by "api_dir" property.

C++ and Java shared libraries are created from the frozen API dumps in
the version directories as well as from the current source code.

To link against the previous version of the interface, use
<modulename>-V<version>-cpp for C++ and <modulename>-V<version>-java for
Java.

To freeze the current AIDL source as a new version, create a new directory
for the version, append the new version name to the 'versions' property,
and invoke 'm <modulename>-freeze-api'.

Bug: 116037747
Bug: 113550618

Test: m
Test: modify build/test_package/IBaz.aidl and run m
=> nothing breaks (since it isn't frozen yet)

Test: modify build/test_package/IBaz.aidl, mkdir build/api/3, add "3" to
versions property of test-root-package, and then run
m test-root-package-freeze-api
=> new API dump is created under build/api/3

Test: modify build/api/2/IBaz.aidl in a backward incompatible way (i.e.
delete a method), and then run m
=> build breaks

Test: mma under build
=> test-[root|piece-1|piece-2|piece-3|piece-4]-V1-java.jar under $(TARGET_OUT)/system/framework and
   test-[root|piece-1|piece-2|piece-3|piece-4]-V1-cpp.so under $(TARGET_OUT)/system/lib

Change-Id: Ib622090696114e75505cc05fe9348e7e264c2c51