Fix circular dependency when aidl_interface is used by framework.jar
Circular dependency can happen when a java library from an
aidl_interface is used for compiling framework.jar. This is because the
java library is built with sdk_version:"current"; it is linked against
the stubs library created from the current framework source code and to
create the stubs library framework.jar is needed.

Temporarily fixing the issue by building the java library with
sdk_version:"28" to cut the dependency to framework.jar. We now build
against the prebuilt stubs and thus we don't need to build framework.jar
to have the stubs lib.

Bug: 116761447
Test: add test-root-package-java to libs property of framework-defaults
in frameworks/base/Android.bp and run 'm framework'.

Change-Id: Id9310789e0d23ea5571f765668be6c8afa1de6ce