The java lib for aidl_interface is built with Java 1.8
The java lib for aidl_interface is building with sdk_version: "28"
to cut the circular dependency to framework.jar
(9435341c3e41961ad19e4639c71c7628b62e3cc9)

However, this prevents the java lib to be built with Java 1.9 because
we don't have system modules defined for the prebuilt stub libs for the
versioned sdk.

Until the issue is fixed, the java lib for aidl_interface is forced to
be built with Java 1.8. (b/117069453)

Bug: 116761447
Test: EXPERIMENTAL_USE_OPENJDK9=true make apex_aidl_interface
Change-Id: Iee9e094f04b9301305b1018ac06e5878c670d8e3