Automatic logging of interface method calls
When an AIDL interface is compiled with --lang=cpp and --log option,
information for method calls to the interface is given to the
user-registered callback function.

The information includes name of the method, name and values or the
arguments, duration of the method call, etc. The information is provided
as a JSON object and the callback can be registered to static field
'logFunc' of the proxy class.

Bug: 111163090
Test: m aidl_test_interface-cpp

Change-Id: I6ecc6dfa6b2a63a3e17c0c9190713ce6cb65447a