Fix: not buildable with pre-23 SDK when with ParcelFileDescriptor
This is fixing a regression that when the input AIDL is using
ParcelFileDescriptor, the generated code is not buildable with SDKs
older than 23. While promoting ParcelFileDescriptor to one of the
built-in types, [write|read]TypedObject APIs were used, but they are
introduced with 23 and thus not avaialble in older SDKs.

Fixing the issue by reading/writing ParcelFileDescriptor without using
the APIs but doing the same as the APIs in the generated code.

Bug: 119523810
Test: aidl_unittests
Change-Id: Ie46e262b3f2746cb75e810603cc83b364634d9f0