No NDK library w/ vendor variant.
ndk does not support vendor variants.

Follow-ups:
- b/119771576: resolution to CreateModule defaults should export
      'is vendor' computation from cc.go to here.
- b/121157555: this check should be removed when NDK libraries have a
      separate vendor/platform variant (since we should always have all
      three variants in this case).

This also has a weird implication. If an aidl_interface is
vendor_available, then things depending on it must also be so. If they
aren't vendor_available, then they will have an ndk version which
depends on the ndk version of a library that is being removed here.

Fixes: 121279086
Test: test-vendor (build-time check that vendor_available works)
Change-Id: I6b01354349da5d872e018c9ec2cbc6cdfc9d4a72