Rename LoggableInterface to ILoggableInterface
This fixes re-ran targets when checkbuild is run repeatedly.

aidl_interface expects generation of <src>.h header for any <src> in
srcs. However, aidl generates I*.h for every interface. This is also
embedded in other places like IMPLEMENT_META_INTERFACE. That is, even
though in 'aidl', an interface doesn't have to start with an I, in C++,
it must.

Ideally, we could completely disallow interfaces that don't start with
an 'I'. However, this would break some downstream app builds. So,
instead, in a follow-up, we'll disable it for the Android platform
only.

Bug: 129549755
Test: m aidl_test_loggable_interface && m aidl_test_loggable_interface

Change-Id: Ia0024439d378cea81b60943b03784e3c9548085f