Update build test APIs.
APIs must match ToT so that we know they don't change and we can rely on
them. However, this is for a test, so it doesn't matter in practice.
This moves forward the release process, but we should look at avoiding
needing test APIs frozen.

Bug: 131924172
Test: force checkEquality to be called in aidl_interface.go, repro build
error reported by release team, update APIs to fix error

Change-Id: Iabec6e853655b5f535b35f6e16b53f782f609b24