fix: incorrect output path
This change fixes a bug that when -o option is used without explicitly
specifying the output path, the output path is determined by
concatenating the directory specified by the -o option and the input
file path and the replacing the suffix .aidl to .java.

This especially is problematic in Windows, because the output path then
contains drive letter (e.g. "C:") in the middle of the path if the input
path is an absolute path. For example,

aidl -o C:\output C:\input\IFoo.aidl

resulted the incorrect output path of C:\output\C:\input\IFoo.java.
It should be C:\input\IFoo.java

This change fixes the problem by not concatenating the output dir with
the input file path.

Bug: 134966834
Test: aidl_unittests

Change-Id: I417623089f7409032483a2f6c79e745c0fc340d3