Allow files ending in single line comment w/o \n
Before, if an aidl file ended in "// ..." without a newline, it would
fail to parse. This is because the lexer included a newline in its
definition of single line comments. The error for this wasn't so nice,
so just making it parse.

Although the extra_text concatenation may require additional allocations
here, this is somewhat nicer because it makes keeping track of newlines
orthogonal to other lexing.

Bug: N/A
Test: aidl_unittests
Change-Id: Ic6b16d595d9178f0740f506d4e1f101b6544a8f7