fix: framework-aidl-mapping is working on incomplete set of inputs
This change fixes a bug that framework-aidl-mapping is working on
incomplete set of inputs. Before the other change in the same topic,
the 'srcs' property of this module was set to 'framework-defaults'. This
module then retrieved the 'srcs' property of the framework-defaults
module, and then extracted aidl files from there. However, during the
extraction, this module didn't take sourcefile providing modules (like
filegroup modules being referenced in :<name> syntax) into account. It
just checked whether an element in the srcs property ends with .aidl
suffix and then took the element.

With this change, the srcs property is annotated with `android:"path"`
and is expected to directly point to individual source files and
filegroup modules - without indirectly via framework-defaults module.

This change also added two new properties to get AIDL include paths.

Bug: 139042150
Test: m

Change-Id: I78bdbd802ffecd41347d375e22790727529383cf