getInterfaceVersion() returns -1 on unversioned interface
This change fixes a bug that getInterfaceVersion() on an unversioned
AIDL interface does not work. Now, if the remote side is unversioned
(i.e. getInterfaceVersion() is not defined), it falls back to the
getInterfaceVersion() of the default impl, which usually returns -1 but
can be overridden to return any value that makes sense for the
interface.

Note that in order for the default impl to be used, it must be set via
<IFaceName>.setDefaultImpl().

Bug: 139919934
Test: aidl_unittests
Change-Id: Ia42b44a5d95beb615c7602262364885856cf2148