Disable "VintfStability + cpp + vendor" variant.
This combination implies that we're using an interface which must be
kept stable but without using the copy of libbinder that is on system.
The runtime enforces this, however by disabling this variant at build
time, we can make sure people don't spend a large amount of time
converting writing code using this variant only to realize when testing
later that it cannot work.

Fixes: 149119785
Test: Try to use variant which fails.
Change-Id: I69a65e48d1cc071b1ad0303c7ed8eef6d958a718