Add unstable property to aidl_interface
The property is used to make it explicit that an AIDL interface is not
stable. When it is set to true, the build system doesn't create the API
dumps and check their compatibility/equality.

The default value for the property is false, which means an
aidl_interface module is by default considered stable for safety;
otherwise, someone could mistakenly use an unstable interface across
module/partition boundaries.

In addition, this change requires a stable interface to have the current
API dump even when the `versions` property is empty, which is the case
when the first version of the interface hasn't yet created.

Bug: 152655547
Bug: 152655544
Test: m

Change-Id: Ibf09857d26f0c8bafef8d326c7f64ba3adbeb8cc